refact(eslint): modularize config into separate files
