+++ eslint/general-eslint.config.mjs
@@ -1,0 +1,48 @@
+const generalConfig = {
+  rules: {
+    camelcase: [
+      'error',
+      {
+        ignoreImports: true,
+      },
+    ],
+    'react/jsx-filename-extension': ['error', { extensions: ['.tsx'] }],
+    '@typescript-eslint/consistent-type-imports': [
+      'error',
+      {
+        prefer: 'type-imports',
+      },
+    ],
+    'no-restricted-imports': [
+      'error',
+      {
+        patterns: ['../*'],
+      },
+    ],
+    'react/function-component-definition': [
+      'error',
+      {
+        namedComponents: 'function-declaration',
+        unnamedComponents: 'arrow-function',
+      },
+    ],
+    'no-restricted-syntax': [
+      'error',
+      {
+        selector: 'ExportDefaultDeclaration > FunctionDeclaration',
+        message:
+          'Do not export functions directly in the default. Separate the declaration.',
+      },
+    ],
+    '@typescript-eslint/no-unused-vars': [
+      'error',
+      {
+        args: 'all',
+        argsIgnorePattern: '^_',
+      },
+    ],
+    'func-style': ['error', 'declaration', { allowArrowFunctions: true }],
+  },
+};
+
+export default generalConfig;
