+++ eslint/prettier-eslint.config.mjs
@@ -1,0 +1,15 @@
+import prettier from 'eslint-config-prettier';
+import prettierPlugin from 'eslint-plugin-prettier';
+
+const prettierConfig = {
+  ...prettier,
+  plugins: {
+    prettier: prettierPlugin,
+  },
+  rules: {
+    'prettier/prettier': 'error',
+    quotes: ['error', 'single', { allowTemplateLiterals: false }],
+  },
+};
+
+export default prettierConfig;
